refactor code according to post action
